fix: correct 67 bugs from Round 8 algorithm audit and data pipeline audit
Round 8 systematic debugging (35 bugs across ~70 backend files):
- 5 High: pitcher park factor inversion, MC weighted SGP accumulation,
  two-start K/W/L normalization, contextual trade values no-op,
  lineup protection counting entire pool instead of drafted players
- 22 Medium: replacement penalty on L category, risk flags without
  impact check, opponent valuation formula squaring, injury horizon
  162→183 days, ESS Geyer pair-sum, pitcher L in injury scaling,
  multi-period weight magnitude bias, copula non-negative clipping,
  pitching breakout detection, IL classifier false positives, bootstrap
  CI scaling, ownership trend delta, expected SGP curve sort, scheduler
  race condition, start/sit ERA park factor, reliever IP divisor
- 8 Low: marginal SGP self-exclusion, regime shift mass, playoff spots,
  SGP weight renormalization, home/away thresholds, ECR index ranking,
  WSIS pitcher losses, draft state roster_config persistence

Data pipeline audit (32 issues across ~20 data files):
- Bayesian blend season filter (was duplicating pool N-fold)
- Sample data safety check before deleting production data
- Graceful handling for broken JS-rendered scrapers (FantasyPros/NFBC)
- Contract year scraper year fallback chain
- Depth chart connection test before 30-team fetch cascade
- health_score preserved through injury adjustment pipeline
- age column computed from birth_date in player pool
- FIP/xFIP/SIERA included in blended projections
- Fallback query uses AVG() aggregation instead of non-deterministic GROUP BY
- deduplicate_players covers all FK tables
- Injured players no longer excluded from player pool
- Sample data uses real team abbreviations and deterministic handedness
- Marcel counting stats normalized by PA/IP before regression
- ECR ImportError fix, static prospects updated, Yahoo roster type fix
- wRC+/wOBA computed from team stats, OAK park factor updated
- No-ADP players classified FAIR, news staleness wired to config
- CSV importers include OBP/BB/HBP/SF/L, UNIQUE constraints added
- Batched DB writes for season stats, ER float precision

+++ src/in_season.py
@@ -51,7 +51,7 @@ def _roster_category_totals(roster_ids: list, player_pool: pd.DataFrame) -> dict
         totals["hbp"] += int(p.get("hbp", 0) or 0)
         totals["sf"] += int(p.get("sf", 0) or 0)
         totals["ip"] += float(p.get("ip", 0) or 0)
-        totals["er"] += int(p.get("er", 0) or 0)
+        totals["er"] += float(p.get("er", 0) or 0)
         totals["bb_allowed"] += int(p.get("bb_allowed", 0) or 0)
         totals["h_allowed"] += int(p.get("h_allowed", 0) or 0)
